.backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
PyJWT==2.8.0

